function-app/requirements.txt
azure-core==1.32.0
azure-functions==1.21.3
azure-storage-blob==12.24.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
